+++ diagram-app/main.js
@@ -1,10 +1,8 @@
 import { createEditor } from '@alexguha/rplib-editor';
-import components from './standard_items/components.json' with { type: 'json' };
 
 window.onload = async () => {
     const manager = await createEditor({
-        components,
-        dataSource: './standard_items/abstract_diagrams.txt',
+        dataDir: './data',
     });
     manager.restoreView('test');
 };